pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64", "setuptools_scm>=8"]

[project]
name = "paper-qa-ui"
version = "0.1.0"
description = "Paper-QA with OpenRouter.ai and Ollama - A clean implementation for scientific paper question-answering"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Paper-QA UI Team", email = "hello@futurehouse.org"},
]
maintainers = [
    {name = "Paper-QA UI Team", email = "hello@futurehouse.org"},
]
keywords = ["question answering", "rag", "scientific papers", "openrouter", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.11"
dependencies = [
    "paper-qa[dev]>=5.0.0",
    "python-dotenv>=1.1.1",
    "tenacity>=8.0.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.10.1",
    "pydantic-settings>=2.0.0",
    "gradio>=4.0.0",
    "flake8>=7.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.8.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/Future-House/paper-qa"
Documentation = "https://futurehouse.gitbook.io/futurehouse-cookbook"
Repository = "https://github.com/Future-House/paper-qa"
"Bug Tracker" = "https://github.com/Future-House/paper-qa/issues"

[project.scripts]
paper-qa-cli = "scripts.paper_qa_cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "paper_qa.*",
    "rich.*",
    "httpx.*",
]
ignore_missing_imports = true
